+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ###
 - Fix cross-instance action queue bug
 - Fix unresolved promise bug within action flushing algorithm
+- Fix missing delete action bug
 
 ## 0.1.1 (2021-6-15)
 ### Fixed